backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.42
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==6.2.0
slowapi==0.1.9
structlog==25.4.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-sqlalchemy==0.57b0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
httpx==0.28.1
